requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
aiohttp==3.11.16
pandas==2.3.2
pyarrow==23.0.1
kafka-python==2.3.0
